refactor(fe): co-locate CLI components with their routes, match copy UX
- Move CliHeader and TerminalBlock under routes/(new-styling)/cli/components/
  (only used by the /cli authorize view).
- Move VerifyPanel and CopyCommandRow under
  routes/(new-styling)/manage/(authenticated)/settings/ (only used by the
  settings CliConfirmDialog).
- Drop the now-empty $lib/components/cli/ directory.

Copy UX now matches the existing pattern (ConfirmThisDevice,
SendConfirmationEmail): wrap the button in a Tooltip with manual + hidden
on copy, instead of swapping the icon to a checkmark. waitFor instead of
raw setTimeout.
