+++ src/frontend/src/routes/(new-styling)/cli/components/CliHeader.svelte
