+++ src/frontend/src/routes/(new-styling)/cli/components/TerminalBlock.svelte
