+++ src/frontend/src/routes/(new-styling)/manage/(authenticated)/settings/VerifyPanel.svelte
